Parse Functionality and Scene
Added parse functionality to the build.
The two new script create an object in the cloud, and create a user in
the cloud.

A new Scene has been added that has both two objects, One that links to
the parse cloud, and another for testing purposes that creates objects
and users in the cloud.